requirements.txt
torch
scikit-image
numpy
matplotlib
argparse
pillow
tqdm
IPython